fix get_hw_json.R with replacing \n chars + just saving changes commit
