+++ .gitignore
@@ -2,3 +2,4 @@
 **/.Rhistory
 **/.RData
 **/.Ruserdata
+.Rproj.user
