+++ hwpr/.Rproj.user/CDB5BE05/pcs/windowlayoutstate.pper
@@ -1,9 +1,9 @@
 {
     "left": {
-        "splitterpos": 273,
-        "topwindowstate": "HIDE",
-        "panelheight": 659,
-        "windowheight": 697
+        "splitterpos": 355,
+        "topwindowstate": "NORMAL",
+        "panelheight": 858,
+        "windowheight": 896
     },
     "right": {
         "splitterpos": 320,
