+++ hwpr/.Rproj.user/CDB5BE05/sources/prop/INDEX
@@ -14,6 +14,12 @@
 ~%2Fwlm2023%2Fhwpr%2Fchecker.js="6879CF84"
 ~%2Fwlm2023%2Fhwpr%2Fdarkstyle.scss="ADD055AB"
 ~%2Fwlm2023%2Fhwpr%2Fget_ans.R="E2C99075"
+~%2Fwlm2023%2Fhwpr%2Fget_hw_jsons.R="05B52B35"
+~%2Fwlm2023%2Fhwpr%2Fhandler.js="F782A3B9"
+~%2Fwlm2023%2Fhwpr%2Fhw-hadler.js="D72F19F0"
+~%2Fwlm2023%2Fhwpr%2Fhw1.json="BDE0CE65"
+~%2Fwlm2023%2Fhwpr%2Fhw2.json="9E12201B"
+~%2Fwlm2023%2Fhwpr%2Fhw_handler.js="C1480C27"
 ~%2Fwlm2023%2Fhwpr%2Findex.qmd="F113C920"
 ~%2Fwlm2023%2Fhwpr%2Flevellabs.scss="7326AD68"
 ~%2Fwlm2023%2Fhwpr%2Flightstyle.scss="AC06B306"
