+++ hwpr/.Rproj.user/shared/notebooks/paths
@@ -6,14 +6,14 @@
 /Users/antonangelgardt/wlm2023/hwpr/_lang.yml="BA7B0B31"
 /Users/antonangelgardt/wlm2023/hwpr/_metadata.yml="C01772A8"
 /Users/antonangelgardt/wlm2023/hwpr/_quarto.yml="DEC43DE0"
-/Users/antonangelgardt/wlm2023/hwpr/ans_test.json="63ABFC2A"
-/Users/antonangelgardt/wlm2023/hwpr/check_test.js="4EFE9BF8"
-/Users/antonangelgardt/wlm2023/hwpr/checker.js="0447BEFB"
 /Users/antonangelgardt/wlm2023/hwpr/darkstyle.scss="2520C7B4"
-/Users/antonangelgardt/wlm2023/hwpr/get_ans.R="7982B0AF"
+/Users/antonangelgardt/wlm2023/hwpr/get_hw_jsons.R="4219E22D"
+/Users/antonangelgardt/wlm2023/hwpr/hw1.json="38A25FB4"
+/Users/antonangelgardt/wlm2023/hwpr/hw2.json="D3294E94"
 /Users/antonangelgardt/wlm2023/hwpr/index.qmd="333F4FFE"
 /Users/antonangelgardt/wlm2023/hwpr/levellabs.scss="C2149B13"
 /Users/antonangelgardt/wlm2023/hwpr/lightstyle.scss="F4317418"
 /Users/antonangelgardt/wlm2023/hwpr/part-hw.qmd="BDDDBA00"
+/Users/antonangelgardt/wlm2023/hwpr/test.js="F4341E67"
 /Users/antonangelgardt/wlm2023/hwpr/test.json="5A5593FE"
 /Users/antonangelgardt/wlm2023/hwpr/test.qmd="F2D4FA81"
