+++ hwpr/.quarto/xref/6153fe10
@@ -1,1 +1,1 @@
-{"entries":[],"options":{"chapter-id":"hw1","chapters":true},"headings":["hw1_main","section","section-1","section-2","section-3","section-4","section-5","section-6","section-7","section-8","section-9","hw1_add","section-10","section-11","section-12","section-13","section-14","hw1"]}
+{"options":{"chapter-id":"hw1","chapters":true},"entries":[],"headings":["hw1_main","section","section-1","section-2","section-3","section-4","section-5","section-6","section-7","section-8","section-9","hw1_add","section-10","section-11","section-12","section-13","section-14","hw1"]}
