+++ hwpr/_book/hw1.json
@@ -1,0 +1,1 @@
+hw1_json='[[{"name":"task","hw1-1":"Это задание простого задания `переменная`. 1","hw1-2":"Это задание простого задания. 2","hw1-3":"Это задание простого задания. 3","hw1-4":"Это задание простого задания. 4","hw1-5":"Это задание среднего задания. 5","hw1-6":"Это задание среднего задания. 6","hw1-7":"Это задание среднего задания. 7","hw1-8":"Это задание среднего задания. 8","hw1-9":"Это задание сложного задания. 9","hw1-10":"Это задание сложного задания. 10","hw1-11":"Это задание среднего задания. 11","hw1-12":"Это задание сложного задания. 12","hw1-13":"Это задание сложного задания. 13","hw1-14":"Это задание очень сложного задания. 14","hw1-15":"Это задание очень сложного задания. 15"},{"name":"level","hw1-1":"easy","hw1-2":"easy","hw1-3":"easy","hw1-4":"easy","hw1-5":"medium","hw1-6":"medium","hw1-7":"medium","hw1-8":"medium","hw1-9":"hard","hw1-10":"hard","hw1-11":"medium","hw1-12":"hard","hw1-13":"hard","hw1-14":"extreme","hw1-15":"extreme"},{"name":"input_requirements","hw1-1":"Описание формата инпута. Значение округлите до сотых по математическим правилам. Проверка игнорирует все пробелы. 1","hw1-2":"Описание формата инпута. Значение округлите до сотых по математическим правилам. Проверка игнорирует все пробелы. 2","hw1-3":"Описание формата инпута. Значение округлите до сотых по математическим правилам. Проверка игнорирует все пробелы. 3","hw1-4":"Описание формата инпута. Значение округлите до сотых по математическим правилам. Проверка игнорирует все пробелы. 4","hw1-5":"Описание формата инпута. Значение округлите до сотых по математическим правилам. Проверка игнорирует все пробелы. 5","hw1-6":"Описание формата инпута. Значение округлите до сотых по математическим правилам. Проверка игнорирует все пробелы. 6","hw1-7":"Описание формата инпута. Значение округлите до сотых по математическим правилам. Проверка игнорирует все пробелы. 7","hw1-8":"Описание формата инпута. Значение округлите до сотых по математическим правилам. Проверка игнорирует все пробелы. 8","hw1-9":"Описание формата инпута. Значение округлите до сотых по математическим правилам. Проверка игнорирует все пробелы. 9","hw1-10":"Описание формата инпута. Значение округлите до сотых по математическим правилам. Проверка игнорирует все пробелы. 10","hw1-11":"Описание формата инпута. Значение округлите до сотых по математическим правилам. Проверка игнорирует все пробелы. 11","hw1-12":"Описание формата инпута. Значение округлите до сотых по математическим правилам. Проверка игнорирует все пробелы. 12","hw1-13":"Описание формата инпута. Значение округлите до сотых по математическим правилам. Проверка игнорирует все пробелы. 13","hw1-14":"Описание формата инпута. Значение округлите до сотых по математическим правилам. Проверка игнорирует все пробелы. 14","hw1-15":"Описание формата инпута. Значение округлите до сотых по математическим правилам. Проверка игнорирует все пробелы. 15"},{"name":"hint_titles","hw1-1":"Как делать задание?||Что надо сделать?||Не принимает ответ","hw1-2":"Как делать задание?||Что надо сделать?","hw1-3":"Как делать задание?||Что надо сделать?","hw1-4":"Как делать задание?||Что надо сделать?","hw1-5":"Как делать задание?||Что надо сделать?","hw1-6":"Как делать задание?||Что надо сделать?","hw1-7":"Как делать задание?||Что надо сделать?","hw1-8":"Как делать задание?||Что надо сделать?","hw1-9":"Как делать задание?||Что надо сделать?","hw1-10":"Как делать задание?||Что надо сделать?","hw1-11":"Как делать задание?||Что надо сделать?","hw1-12":"Как делать задание?||Что надо сделать?","hw1-13":"Как делать задание?||Что надо сделать?","hw1-14":"Как делать задание?||Что надо сделать?","hw1-15":"Как делать задание?||Что надо сделать?"},{"name":"hints","hw1-1":"Тест подсказки 1-1.||Текст подсказки 1-2.||Тест подсказки 1-3.","hw1-2":"Тест подсказки 2-1.||Текст подсказки 2-2.","hw1-3":"Тест подсказки 3-1.||Текст подсказки 3-2.","hw1-4":"Тест подсказки 4-1.||Текст подсказки 4-2.","hw1-5":"Тест подсказки 5-1.||Текст подсказки 5-2.","hw1-6":"Тест подсказки 6-1.||Текст подсказки 6-2.","hw1-7":"Тест подсказки 7-1.||Текст подсказки 7-2.","hw1-8":"Тест подсказки 8-1.||Текст подсказки 8-2.","hw1-9":"Тест подсказки 9-1.||Текст подсказки 9-2.","hw1-10":"Тест подсказки 10-1.||Текст подсказки 10-2.","hw1-11":"Тест подсказки 11-1.||Текст подсказки 11-2.","hw1-12":"Тест подсказки 12-1.||Текст подсказки 12-2.","hw1-13":"Тест подсказки 13-1.||Текст подсказки 13-2.","hw1-14":"Тест подсказки 14-1.||Текст подсказки 14-2.","hw1-15":"Тест подсказки 15-1.||Текст подсказки 15-2."},{"name":"has_autocheck","hw1-1":"True","hw1-2":"True","hw1-3":"True","hw1-4":"True","hw1-5":"True","hw1-6":"True","hw1-7":"True","hw1-8":"True","hw1-9":"True","hw1-10":"True","hw1-11":"True","hw1-12":"True","hw1-13":"True","hw1-14":"True","hw1-15":"False"},{"name":"autocheck_answer","hw1-1":"1","hw1-2":"2","hw1-3":"3","hw1-4":"4","hw1-5":"5","hw1-6":"6","hw1-7":"7","hw1-8":"8","hw1-9":"9","hw1-10":"10","hw1-11":"11","hw1-12":"12","hw1-13":"13","hw1-14":"14","hw1-15":""}]]'
