+++ hwpr/_book/test.html
@@ -183,7 +183,6 @@ ul.task-list li input[type="checkbox"] {
 
 <!--
 REPLACE ALL hw1 TO hw<n> WITH MATCH CASE
-CHANGE test to hw<n> IN THE SCRIPT CHUNCK IN THE END OF THE FILE
 -->
 <section id="hw1_main" class="level2">
 <h2 class="anchored" data-anchor-id="hw1_main">Основные задания</h2>
@@ -519,9 +518,10 @@ CHANGE test to hw<n> IN THE SCRIPT CHUNCK IN THE END OF THE FILE
 Подсказки
 </summary>
 </details>
-<!--- CHANGE test TO hw<n> IN THE LINES BELOW--->
-<script type="text/javascript" src="test.json"></script>
-<script type="module" src="test.js"></script>
+<!--- CHANGE hw1 TO hw<n> IN THE LINES BELOW--->
+<script type="text/javascript" src="hw1.json"></script>
+<script type="module" src="handler.js"></script>
+<script type="module" src="hw_handler.js"></script>
 
 
 </section>
