+++ hwpr/hw2.json
@@ -1,0 +1,1 @@
+hw2_json='[[{"name":"task","hw2-1":"","hw2-2":"","hw2-3":"","hw2-4":"","hw2-5":"","hw2-6":"","hw2-7":"","hw2-8":"","hw2-9":"","hw2-10":"","hw2-11":"","hw2-12":"","hw2-13":"","hw2-14":"","hw2-15":""},{"name":"level","hw2-1":"","hw2-2":"","hw2-3":"","hw2-4":"","hw2-5":"","hw2-6":"","hw2-7":"","hw2-8":"","hw2-9":"","hw2-10":"","hw2-11":"","hw2-12":"","hw2-13":"","hw2-14":"","hw2-15":""},{"name":"input_requirements","hw2-1":"","hw2-2":"","hw2-3":"","hw2-4":"","hw2-5":"","hw2-6":"","hw2-7":"","hw2-8":"","hw2-9":"","hw2-10":"","hw2-11":"","hw2-12":"","hw2-13":"","hw2-14":"","hw2-15":""},{"name":"hint_titles","hw2-1":"","hw2-2":"","hw2-3":"","hw2-4":"","hw2-5":"","hw2-6":"","hw2-7":"","hw2-8":"","hw2-9":"","hw2-10":"","hw2-11":"","hw2-12":"","hw2-13":"","hw2-14":"","hw2-15":""},{"name":"hints","hw2-1":"","hw2-2":"","hw2-3":"","hw2-4":"","hw2-5":"","hw2-6":"","hw2-7":"","hw2-8":"","hw2-9":"","hw2-10":"","hw2-11":"","hw2-12":"","hw2-13":"","hw2-14":"","hw2-15":""},{"name":"has_autocheck","hw2-1":"True","hw2-2":"True","hw2-3":"True","hw2-4":"True","hw2-5":"True","hw2-6":"True","hw2-7":"True","hw2-8":"True","hw2-9":"True","hw2-10":"True","hw2-11":"True","hw2-12":"True","hw2-13":"True","hw2-14":"True","hw2-15":"True"},{"name":"autocheck_answer","hw2-1":"16","hw2-2":"17","hw2-3":"18","hw2-4":"19","hw2-5":"20","hw2-6":"21","hw2-7":"22","hw2-8":"23","hw2-9":"24","hw2-10":"25","hw2-11":"26","hw2-12":"27","hw2-13":"28","hw2-14":"29","hw2-15":"30"}]]'
